Added WindowComparator finder to ResistorCalcs.py
Added a function to find the closest real resistor values for building a window comparator for OV and UV alerts.  Given a Series (E24, E96, etc), will start by finding the required ratios between the resistors in the string, find the closest value available for the Middle resistor, find the closest available Bottom resistor, and then split the required Top resistance into two units for greater precision.  It may be that the next larger value for Top was closer than the next smaller value; however, by choosing the lower value and adding a second series resistance the precision will nearly always be more exact than any single value.   If the additional resistor is very small indeed, it can of course be ignored.